+++ cpp/src/snmg/COO2CSR/COO2CSR.cu
@@ -290,13 +290,19 @@ gdf_error snmg_coo2csr_impl(size_t* part_offsets,
 
 #pragma omp barrier
 
+  int myRowCount = 0;
+  for (int j = 0; j < p; j++){
+    idx_t* otherRowCounts = comm->rowCounts + (j * p);
+    myRowCount += otherRowCounts[i];
+  }
+
   // Each thread allocates space to receive their rows from others
   idx_t *cooRowNew, *cooColNew;
   val_t *cooValNew;
-  ALLOC_TRY(&cooRowNew, sizeof(idx_t) * myEdgeCount, nullptr);
-  ALLOC_TRY(&cooColNew, sizeof(idx_t) * myEdgeCount, nullptr);
+  ALLOC_TRY(&cooRowNew, sizeof(idx_t) * myRowCount, nullptr);
+  ALLOC_TRY(&cooColNew, sizeof(idx_t) * myRowCount, nullptr);
   if (cooValTemp != nullptr) {
-    ALLOC_TRY(&cooValNew, sizeof(val_t) * myEdgeCount, nullptr);
+    ALLOC_TRY(&cooValNew, sizeof(val_t) * myRowCount, nullptr);
   }
   else {
     cooValNew = nullptr;
@@ -304,6 +310,8 @@ gdf_error snmg_coo2csr_impl(size_t* part_offsets,
   comm->rowPtrs[i] = cooRowNew;
   comm->colPtrs[i] = cooColNew;
   comm->valPtrs[i] = cooValNew;
+  cudaCheckError();
+  cudaDeviceSynchronize();
 #pragma omp barrier
 
   // Each thread copies the rows needed by other threads to them
@@ -314,22 +322,22 @@ gdf_error snmg_coo2csr_impl(size_t* part_offsets,
       idx_t* prevRowCounts = comm->rowCounts + (prev * p);
       offset += prevRowCounts[other];
     }
-    cudaMemcpyPeer(comm->rowPtrs[other] + offset,
-                   other,
-                   cooRowTemp + positions[other],
-                   i,
-                   rowCount * sizeof(idx_t));
-    cudaMemcpyPeer(comm->colPtrs[other] + offset,
-                   other,
-                   cooColTemp + positions[other],
-                   i,
-                   rowCount * sizeof(idx_t));
-    if (cooValTemp != nullptr) {
-      cudaMemcpyPeer(comm->valPtrs[other] + offset,
-                     other,
-                     cooValTemp + positions[other],
-                     i,
-                     rowCount * sizeof(idx_t));
+
+    if (rowCount > 0) {
+      cudaMemcpy(comm->rowPtrs[other] + offset,
+                 cooRowTemp + positions[other],
+                 rowCount * sizeof(idx_t),
+                 cudaMemcpyDefault);
+      cudaMemcpy(comm->colPtrs[other] + offset,
+                 cooColTemp + positions[other],
+                 rowCount * sizeof(idx_t),
+                 cudaMemcpyDefault);
+      if (cooValTemp != nullptr) {
+        cudaMemcpy(comm->valPtrs[other],
+                   cooValTemp + positions[other],
+                   rowCount * sizeof(idx_t),
+                   cudaMemcpyDefault);
+      }
     }
   }
   cudaCheckError();
@@ -353,7 +361,7 @@ gdf_error snmg_coo2csr_impl(size_t* part_offsets,
   idx_t myOffset = part_offsets[i];
   thrust::transform(rmm::exec_policy(nullptr)->on(nullptr),
                     cooRowNew,
-                    cooRowNew + myEdgeCount,
+                    cooRowNew + myRowCount,
                     thrust::make_constant_iterator(myOffset * -1),
                     cooRowNew,
                     thrust::plus<idx_t>());
@@ -363,7 +371,7 @@ gdf_error snmg_coo2csr_impl(size_t* part_offsets,
     auto zippy = thrust::make_zip_iterator(thrust::make_tuple(cooRowNew, cooColNew));
     thrust::sort_by_key(rmm::exec_policy(nullptr)->on(nullptr),
                         zippy,
-                        zippy + myEdgeCount,
+                        zippy + myRowCount,
                         cooValNew);
   }
   else {
@@ -389,7 +397,7 @@ gdf_error snmg_coo2csr_impl(size_t* part_offsets,
                                      unique,
                                      counts,
                                      runcount,
-                                     myEdgeCount);
+                                     myRowCount);
   ALLOC_TRY(&tmpStorage, tmpBytes, nullptr);
   cub::DeviceRunLengthEncode::Encode(tmpStorage,
                                      tmpBytes,
@@ -397,7 +405,7 @@ gdf_error snmg_coo2csr_impl(size_t* part_offsets,
                                      unique,
                                      counts,
                                      runcount,
-                                     myEdgeCount);
+                                     myRowCount);
   ALLOC_FREE_TRY(tmpStorage, nullptr);
 
   cudaDeviceSynchronize();
@@ -426,15 +434,14 @@ gdf_error snmg_coo2csr_impl(size_t* part_offsets,
   csrOff->dtype = cooRow->dtype;
   csrOff->size = localMaxId + 2;
   csrOff->data = offsets;
-
   cugraph::gdf_col_set_defaults(csrInd);
   csrInd->dtype = cooRow->dtype;
-  csrInd->size = myEdgeCount;
+  csrInd->size = myRowCount;
   csrInd->data = cooColNew;
   if (cooValNew != nullptr) {
     cugraph::gdf_col_set_defaults(cooVal);
     csrVal->dtype = cooVal->dtype;
-    csrVal->size = myEdgeCount;
+    csrVal->size = myRowCount;
     csrVal->data = cooValNew;
   }
 #pragma omp barrier