+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 - PR #364 Fixed bug building or installing cugraph when conda isn't installed
 - PR #375 Added a function to initialize gdf columns in cugraph #375
 - PR #378 cugraph was unable to import device_of_gpu_pointer
+- PR #384 Fixed bug in snmg coo2csr causing error in dask-cugraph tests.
 - PR #382 Disabled vertex id check to allow Azure deployment